Fix some float to double type problems.

+++ moments.h
@@ -104,7 +104,7 @@ double momLocrAddMomr5cm(LOCR *l,MOMR *m,momFloat dir,momFloat x,momFloat y,momF
 double momLocrAddMono5(LOCR *l,momFloat m,momFloat dir,momFloat x,momFloat y,momFloat z,double *tax,double *tay,double *taz);
 double momFlocrAddFmomr5cm(FLOCR *l,float v,FMOMR *m,float u,float dir,float x,float y,float z,float *tax,float *tay,float *taz);
 double momFlocrAddMono5(FLOCR *l,float v,float m,float dir,float x,float y,float z,float *tax,float *tay,float *taz);
-double momLocrAddFmomr5cm(LOCR *l,FMOMR *m,float u,float dir,float x,float y,float z,float *tax,float *tay,float *taz);
+double momLocrAddFmomr5cm(LOCR *l,FMOMR *m,double u,double dir,double x,double y,double z,double *tax,double *tay,double *taz);
 /*
 ** All the functions for evaluating local expansions.
 */
